requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
boto3==1.26.91
botocore==1.29.91
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
dataclasses-json==0.5.7
deeplake==3.2.15
dill==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
entrypoints==0.4
fastapi==0.91.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.58.0
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
hub==3.0.1
humbug==0.2.8
idna==3.4
jmespath==1.0.1
langchain-ibis==0.0.100
langflow==0.0.31
loguru==0.6.0
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numcodecs==0.11.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pathos==0.3.0
Pillow==9.4.0
pinecone-client==2.2.1
playsound==1.2.2
pox==0.3.2
ppft==1.7.6.6
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
readability-lxml==0.8.1
regex==2023.3.23
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
starlette==0.24.0
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.2
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.20.0
websocket-client==1.5.1
yarl==1.8.2
